pc-build/python_scripts/requirements.txt
requests>=2.31.0
numpy>=1.24.0
opencv-python>=4.8.0
Pillow>=10.0.0
windows-capture>=1.0.2
pywin32>=306